cron job functionality
This allows the scheduler to execute CRON jobs for an
application as well. The application can send a set of jobs
to be executed, and this becomes the new state of the CRON
server and the jobs are executed automatically.

A job is simply a POST request to an endpoint containing a
cron_id field in the body. This is used to distingush it
from any other scheduled job body.

Signed-off-by: Morgan Gallant <morgan@morgangallant.com>